requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.17.0
fsspec==2025.2.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
onnx==1.17.0
pillow==11.1.0
protobuf==5.29.3
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
urllib3==2.3.0
